Refactor: Snap State and Account Static Classes (#111)
