+++ packages/hedera-wallet-snap/packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "git+https://github.com/hashgraph/hedera-metamask-snaps.git"
   },
   "source": {
-    "shasum": "bg0dUT+K57KjvAI8AB1sQAdW/r5QSG/XamegqcfZxhc=",
+    "shasum": "aJWBuxTQqeVNei+Y/kY9Wp0+HbXJ27a6I5jMpAL7ItI=",
     "location": {
       "npm": {
         "filePath": "dist/snap.js",
