+++ packages/hedera-wallet-snap/packages/snap/src/index.ts
@@ -26,10 +26,9 @@ import _ from 'lodash';
 import { getAccountBalance } from './rpc/account/getAccountBalance';
 import { getAccountInfo } from './rpc/account/getAccountInfo';
 import { transferCrypto } from './rpc/transactions/transferCrypto';
-import { setCurrentAccount } from './snap/account';
-import { getSnapStateUnchecked, initSnapState } from './snap/state';
+import { SnapAccounts } from './snap/SnapAccounts';
+import { SnapState } from './snap/SnapState';
 import { WalletSnapParams } from './types/state';
-
 import { approveAllowance } from './rpc/account/approveAllowance';
 import { deleteAccount } from './rpc/account/deleteAccount';
 import { deleteAllowance } from './rpc/account/deleteAllowance';
@@ -38,8 +37,8 @@ import { associateTokens } from './rpc/hts/associateTokens';
 import { createToken } from './rpc/hts/createToken';
 import { signMessage } from './rpc/misc/signMessage';
 import { getTransactions } from './rpc/transactions/getTransactions';
-import { StakeHbarRequestParams } from './types/params';
 import { HederaUtils } from './utils/HederaUtils';
+import { StakeHbarRequestParams } from './types/params';
 
 /**
  * Handle incoming JSON-RPC requests, sent through `wallet_invokeSnap`.
@@ -63,9 +62,9 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
     JSON.stringify(request.params, null, 4),
   );
 
-  let state = await getSnapStateUnchecked();
+  let state = await SnapState.getStateUnchecked();
   if (_.isEmpty(state)) {
-    state = await initSnapState();
+    state = await SnapState.initState();
   }
 
   let isExternalAccount = false;
@@ -79,7 +78,7 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
   );
 
   // Set current account
-  await setCurrentAccount(
+  await SnapAccounts.setCurrentAccount(
     origin,
     state,
     request.params,
